mythccextractor: rename commandlineparser.* to mythccextractor_...
